// frontend/src/models/Villager.ts
export interface Villager {
  name: string;
  iconImage: string;
  photoImage: string;
  species: string;
  gender: string;
  personality: string;
  hobby: string;
  birthday: string;
  catchphrase: string;
  favoriteSong: string;
  favoriteSaying: string;
  style1: string;
  style2: string;
  color1: string;
  color2: string;
  nameColor: string;
  bubbleColor: string;
  id: string;
}